rename golde model py
